chore: linter unification
- [x] Use extended config versions of eslint and tsconfig
