+++ core/tsconfig.json
@@ -1,34 +1,7 @@
 {
   "extends": "../tsconfig.json",
   "compilerOptions": {
-    "outDir": "./build",
-    "declaration": true,
-    "allowJs": true,
-    "noUnusedLocals": true,
-    "noUnusedParameters": true,
-    "noImplicitReturns": true,
-    "target": "ES2020",
-    "lib": ["es2020"],
-    "experimentalDecorators": true,
-    "strict": true,
-    "noFallthroughCasesInSwitch": true,
-    "noImplicitAny": true,
-    "strictPropertyInitialization": true,
-    "strictNullChecks": true,
-    "module": "commonjs",
-    "moduleResolution": "node",
-    "resolveJsonModule": true,
-    "skipLibCheck": true,
-    "esModuleInterop": true,
-    "isolatedModules": true,
-    "forceConsistentCasingInFileNames": true,
-    "allowSyntheticDefaultImports": true,
-    "composite": true,
-    "incremental": true,
-    "declarationMap": true,
-    "sourceMap": true,
-    "stripInternal": true
+    "outDir": "./build"
   },
-  "include": ["./ts"],
-  "exclude": ["./ts/__tests__.old"]
+  "include": ["./ts"]
 }
