+++ crypto/.eslintrc.js
@@ -1,44 +1,9 @@
-const fs = require("fs");
 const path = require("path");
-
-const prettierConfig = fs.readFileSync(path.resolve(__dirname, "../.prettierrc"), "utf8");
-const prettierOptions = JSON.parse(prettierConfig);
-const isProduction = process.env.NODE_ENV === "production";
 
 module.exports = {
   root: true,
-  extends: [
-    "airbnb",
-    "prettier",
-    "plugin:import/recommended",
-    "plugin:@typescript-eslint/eslint-recommended",
-    "plugin:@typescript-eslint/recommended",
-    "plugin:@typescript-eslint/recommended-type-checked",
-    "plugin:@typescript-eslint/strict",
-    "plugin:@typescript-eslint/strict-type-checked",
-    "plugin:@typescript-eslint/stylistic",
-    "plugin:@typescript-eslint/stylistic-type-checked",
-    "plugin:import/typescript",
-  ],
-  plugins: ["json", "prettier", "unused-imports", "import", "@typescript-eslint"],
+  extends: ["../.eslintrc.js"],
   parser: "@typescript-eslint/parser",
-  env: {
-    node: true,
-    mocha: true,
-    es2022: true,
-  },
-  settings: {
-    react: {
-      version: "999.999.999",
-    },
-    "import/resolver": {
-      typescript: {},
-      node: {
-        extensions: [".ts", ".js"],
-        moduleDirectory: ["node_modules", "ts", "src"],
-      },
-    },
-  },
   parserOptions: {
     project: path.resolve(__dirname, "./tsconfig.json"),
     sourceType: "module",
@@ -51,57 +16,5 @@ module.exports = {
       impliedStrict: true,
     },
     warnOnUnsupportedTypeScriptVersion: true,
-  },
-  reportUnusedDisableDirectives: isProduction,
-  rules: {
-    "import/no-cycle": ["error"],
-    "unused-imports/no-unused-imports": "error",
-    "import/no-extraneous-dependencies": [
-      "error",
-      {
-        devDependencies: ["**/*.test.ts"],
-      },
-    ],
-    "no-debugger": isProduction ? "error" : "off",
-    "no-console": "error",
-    "no-underscore-dangle": "error",
-    "no-redeclare": ["error", { builtinGlobals: true }],
-    "import/order": [
-      "error",
-      {
-        groups: ["external", "builtin", "internal", "type", "parent", "sibling", "index", "object"],
-        alphabetize: {
-          order: "asc",
-          caseInsensitive: true,
-        },
-        warnOnUnassignedImports: true,
-        "newlines-between": "always",
-      },
-    ],
-    "prettier/prettier": ["error", prettierOptions],
-    "import/prefer-default-export": "off",
-    "import/extensions": ["error", { json: "always" }],
-    "class-methods-use-this": "off",
-    "prefer-promise-reject-errors": "off",
-    "max-classes-per-file": "off",
-    "no-use-before-define": ["off"],
-    "no-shadow": "off",
-    curly: ["error", "all"],
-
-    "@typescript-eslint/explicit-member-accessibility": ["error", { accessibility: "no-public" }],
-    "@typescript-eslint/no-non-null-assertion": "off",
-    "@typescript-eslint/prefer-nullish-coalescing": "off",
-    "@typescript-eslint/no-floating-promises": "off",
-    "@typescript-eslint/no-explicit-any": "error",
-    "@typescript-eslint/explicit-module-boundary-types": "error",
-    "@typescript-eslint/no-use-before-define": ["error", { functions: false, classes: false }],
-    "@typescript-eslint/no-misused-promises": ["error", { checksVoidReturn: false }],
-    "@typescript-eslint/no-shadow": [
-      "error",
-      {
-        builtinGlobals: true,
-        allow: ["location", "event", "history", "name", "status", "Option", "test", "expect"],
-      },
-    ],
   },
 };
