+++ tsconfig.json
@@ -1,25 +1,34 @@
 {
   "compilerOptions": {
-    "target": "ES2020",
-    "lib": ["es2020"],
     "module": "commonjs",
     "moduleResolution": "node",
-    "experimentalDecorators": true,
     "alwaysStrict": false,
-    "allowJs": true,
-    "noImplicitAny": false,
     "forceConsistentCasingInFileNames": true,
-    "noUnusedLocals": false,
-    "noUnusedParameters": false,
-    "noImplicitReturns": true,
     "noFallthroughCasesInSwitch": true,
     "sourceMap": true,
-    "strict": false,
+    "outDir": "./build",
+    "declaration": true,
+    "allowJs": true,
+    "noUnusedLocals": true,
+    "noUnusedParameters": true,
+    "noImplicitReturns": true,
+    "target": "ES2020",
+    "lib": ["es2020"],
+    "experimentalDecorators": true,
+    "strict": true,
+    "noImplicitAny": true,
+    "strictPropertyInitialization": true,
+    "strictNullChecks": true,
+    "resolveJsonModule": true,
     "skipLibCheck": true,
     "esModuleInterop": true,
-    "outDir": "./build",
-    "declaration": true
+    "isolatedModules": true,
+    "allowSyntheticDefaultImports": true,
+    "composite": true,
+    "incremental": true,
+    "declarationMap": true,
+    "stripInternal": true
   },
   "exclude": ["node_modules"],
-  "include": ["./**/ts"]
+  "include": ["./**/ts", "./**/typechain-types"]
 }
